Using next/Head breaks things, so does an empty layout
