+++ src/app/articles/head.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+// noinspection JSUnusedGlobalSymbols
+export default function Head(): React.ReactNode {
+  return (
+    <head>
+      <title>Articles - aylett.co.uk</title>
+      <link rel="icon" href="/favicon.ico" />
+    </head>
+  );
+}
