+++ src/app/links/head.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+// noinspection JSUnusedGlobalSymbols
+export default function Head(): React.ReactNode {
+  return (
+    <head>
+      <title>Links</title>
+      <link rel="icon" href="/favicon.ico" />
+    </head>
+  );
+}
