+++ src/app/links/page.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 
-import Head from 'next/head';
 import Link from 'next/link';
 import 'server-only';
 
@@ -10,10 +9,6 @@ import styles from '../../index.module.css';
 export default function Links(): React.ReactNode {
   return (
     <div className={styles.container}>
-      <Head>
-        <title>Links</title>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
       <main>
         <nav>
           <Link href="/">Home</Link>
